requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0.post1
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
greenlet==3.2.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart>=0.0.7
httpx==0.25.2
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
email-validator==2.1.0
reportlab==4.0.8
python-barcode==0.15.1
qrcode==7.4.2
Pillow>=10.4.0
faker==20.1.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.0
openai==1.12.0
strawberry-graphql[fastapi]>=0.288.0
docxtpl>=0.16.0
openpyxl>=3.1.0
weasyprint>=60.0
jinja2>=3.1.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-bdd==6.1.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0